refactor(engineer_loop): split tests_mod.rs (#1266)
tests_mod.rs 1130 -> 381 LOC; new tests_mod_more.rs (~378) + tests_mod_most.rs (~389).
